[Responsive] Deprecate Responsive component in favour of Media component.
