+++ src/Apps/Artist/Routes/AuctionResults/TableSidebar.tsx
@@ -2,7 +2,7 @@ import { Box, Flex, LargeSelect, Sans, Separator, Spacer } from "@artsy/palette"
 import React from "react"
 import { Col, Row } from "Styleguide/Elements/Grid"
 import { Subscribe } from "unstated"
-import { Media } from "Utils/Responsive/Media2"
+import { Media } from "Utils/Responsive"
 import { AuctionResultsState } from "./state"
 
 const SORTS = [
